backend/python/ocr/requirements.txt
grpcio==1.71.0
boto3==1.38.23
rich==14.0.0
protobuf==6.31.0
